Added an example Feature for the project.
I added these example feature as I am new to python projects and I wanted to add something to your project. The .Py file I added is an example of a timer clock that counts the amount of seconds passed. I thought this would be an interesting feature to add to your file as it would allow users to track the actually time difference between the various Algorithms.